Add optional binding to input decorator docs
As mentioned in #139, I made it more clear that inputs were optional. There is already a note in the description of the decorator parameter but I added it to the top as well. I didn't put the optional symbol in the table because the table is mapping it to the decorator parameters (`[property]` => `@Input("<")`)